PLAT-14774 Disable/Override custom add to cart button shortcode based on showAddToCart option value
In this commit we are disabling/enabling the custom add to cart button short-code based on the showAddToCart global settings option value.
If ON the short-code will render that add to cart button if OFF the short-code will return an empty string and no button will be visible on the page.